Make properties unsettable, fix placeholder
