+++ graphql.schema.json
@@ -9068,6 +9068,20 @@
                 "defaultValue": null
               },
               {
+                "name": "claimType",
+                "description": null,
+                "type": {
+                  "kind": "NON_NULL",
+                  "name": null,
+                  "ofType": {
+                    "kind": "SCALAR",
+                    "name": "String",
+                    "ofType": null
+                  }
+                },
+                "defaultValue": null
+              },
+              {
                 "name": "propertyId",
                 "description": null,
                 "type": {
