+++ src/features/claims/claim-details/components/ClaimType/graphql/unset-claim-property-selection.graphql
@@ -1,0 +1,24 @@
+mutation UnsetClaimPropertySelection(
+  $id: ID!
+  $claimType: String!
+  $propertyId: ID!
+) {
+  unsetClaimPropertySelection(
+    id: $id
+    claimType: $claimType
+    propertyId: $propertyId
+  ) {
+    id
+    propertySelections {
+      claimType
+      property {
+        id
+        name
+      }
+      option {
+        id
+        name
+      }
+    }
+  }
+}
